Squash all the commits containing Wago footprint
This should get rid of any content containing the
copyrighted footprint even on the board. Not
sure if this is really needed but I want to avoid
any problems with licensing.

This commit squashes all these commits:
---
Initial commit

---

Finish layout

----

Add TVS for I2C interface

---

Update component values and add clamping diode

The input resistor for GND lvl measurement was above the
recommended value for ADS1015.

The extra clamping diode is needed in case the board is unpowered
and GND lvl is connected to a positive voltage to push +5V to
above GND lvl.

---

Switch mountings to M3 and update I2C conn

M3 matches the srews used for the shut resistors.

The I2C conn now matches the Adafruit Stemma or
the Grove connector.

This required a lot of shiftig stuff around...

---

Replace WAGO footprint with hand creatd one

This should avoid copyright problems. Also in my
opinion it looks better ;)